transfer nft and transfer nft to did
