+++ crates/chia-sdk-driver/src/puzzles/nft/nft_spend.rs
@@ -1,79 +1,3 @@
-use chia_protocol::{Bytes32, CoinSpend};
-use chia_puzzles::nft::{
-    NftOwnershipLayerArgs, NftOwnershipLayerSolution, NftRoyaltyTransferPuzzleArgs,
-    NftStateLayerArgs, NftStateLayerSolution,
-};
-use chia_sdk_types::{
-    conditions::{Condition, NewNftOwner},
-    puzzles::NftInfo,
-};
-use clvm_traits::{clvm_list, ToClvm, ToNodePtr};
-use clvm_utils::{tree_hash, CurriedProgram};
-use clvmr::{
-    sha2::{Digest, Sha256},
-    Allocator, NodePtr,
-};
-
-use crate::{spend_singleton, Conditions, Spend, SpendContext, SpendError};
-
-#[derive(Debug, Clone)]
-pub struct TransferNft<M> {
-    pub did_conditions: Conditions,
-    pub p2_conditions: Conditions,
-    pub output: NftInfo<M>,
-}
-
-pub fn transfer_nft<M>(
-    ctx: &mut SpendContext,
-    nft_info: &NftInfo<M>,
-    p2_puzzle_hash: Bytes32,
-    new_nft_owner: Option<NewNftOwner>,
-) -> Result<TransferNft<M>, SpendError>
-where
-    M: ToClvm<NodePtr> + Clone,
-{
-    let mut did_conditions = Conditions::new();
-    let mut p2_conditions =
-        Conditions::new().create_hinted_coin(p2_puzzle_hash, nft_info.coin.amount, p2_puzzle_hash);
-
-    let new_owner = if let Some(new_nft_owner) = new_nft_owner {
-        did_conditions = did_conditions.assert_raw_puzzle_announcement(did_puzzle_assertion(
-            nft_info.coin.puzzle_hash,
-            &new_nft_owner,
-        ));
-        p2_conditions = p2_conditions.condition(Condition::Other(ctx.alloc(&new_nft_owner)?));
-        new_nft_owner.did_id
-    } else {
-        nft_info.current_owner
-    };
-
-    Ok(TransferNft {
-        did_conditions,
-        p2_conditions,
-        output: nft_info.child(p2_puzzle_hash, new_owner),
-    })
-}
-
-#[allow(clippy::missing_panics_doc)]
-pub fn did_puzzle_assertion(nft_full_puzzle_hash: Bytes32, new_nft_owner: &NewNftOwner) -> Bytes32 {
-    let mut allocator = Allocator::new();
-
-    let new_nft_owner_args = clvm_list!(
-        new_nft_owner.did_id,
-        &new_nft_owner.trade_prices,
-        new_nft_owner.did_inner_puzzle_hash
-    )
-    .to_node_ptr(&mut allocator)
-    .unwrap();
-
-    let mut hasher = Sha256::new();
-    hasher.update(nft_full_puzzle_hash);
-    hasher.update([0xad, 0x4c]);
-    hasher.update(tree_hash(&allocator, new_nft_owner_args));
-
-    Bytes32::new(hasher.finalize().into())
-}
-
 #[cfg(test)]
 mod tests {
     use crate::{nft_mint, IntermediateLauncher, Launcher};
